+++ implementation/src/main/java/com/avioconsulting/mule/testing/muleinterfaces/wrappers/MessageWrapperImpl.groovy
@@ -57,12 +57,8 @@ class MessageWrapperImpl implements
         def klass = value.getClass().name
         // this is what repeatable streams look like
         if (klass.contains('ManagedCursorIteratorProvider')) {
-            def cursor = value.openCursor() as Iterator
-            try {
-                return cursor.toList()
-            }
-            finally {
-                cursor.close()
+            withCursorAsList(value) { List list ->
+                return list
             }
         } else {
             throw new TestingFrameworkException("Do not know how to handle payload of type ${klass}! This method can only be used with iterator/page types")